requirements.txt
Flask 
Flask-Mail
Flask-Cors
python-dotenv